feat: add validations to register and login dtos
- move all user DTOs to dtos/user folder
- add validation dependency at pom
