+++ src/main/java/com/igrowker/wander/controller/AuthController.java
@@ -1,12 +1,13 @@
 package com.igrowker.wander.controller;
 
-import com.igrowker.wander.dto.LoginRequest;
-import com.igrowker.wander.dto.LoginResponse;
-import com.igrowker.wander.dto.RegisterUserDto;
-import com.igrowker.wander.dto.ResponseUserDto;
+import com.igrowker.wander.dto.user.LoginRequest;
+import com.igrowker.wander.dto.user.LoginResponse;
+import com.igrowker.wander.dto.user.RegisterUserDto;
+import com.igrowker.wander.dto.user.ResponseUserDto;
 import com.igrowker.wander.security.JwtService;
 import com.igrowker.wander.service.AuthService;
 import io.swagger.v3.oas.annotations.Operation;
+import jakarta.validation.Valid;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -23,7 +24,7 @@ public class AuthController {
     private JwtService jwtService;
 
     @PostMapping("/register")
-    public ResponseEntity<ResponseUserDto> registerUser(@RequestBody RegisterUserDto userDto) {
+    public ResponseEntity<ResponseUserDto> registerUser(@Valid @RequestBody RegisterUserDto userDto) {
         ResponseUserDto registeredUser = authService.registerUser(userDto);
         return ResponseEntity.status(HttpStatus.CREATED).body(registeredUser);
     }
@@ -40,7 +41,7 @@ public class AuthController {
             description = "This endpoint allows a user to login by providing their credentials. If successful, a JWT token is returned."
     )
     @PostMapping("/login")
-    public ResponseEntity<LoginResponse> login(@RequestBody LoginRequest loginRequest) {
+    public ResponseEntity<LoginResponse> login(@Valid @RequestBody LoginRequest loginRequest) {
         LoginResponse loginResponse = authService.authenticateUser(loginRequest);
         return ResponseEntity.ok(loginResponse);
     }
