--- src/main/java/com/igrowker/wander/dto/RegisterUserDto.java
@@ -1,20 +1,0 @@
-package com.igrowker.wander.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-import java.util.List;
-
-@Data
-@AllArgsConstructor
-@NoArgsConstructor
-public class RegisterUserDto {
-
-    private String name;
-    private String email;
-    private String password;
-    private String role;
-    private List<String> preferences;
-    private String location;
-}
