+++ src/main/java/com/igrowker/wander/dto/user/LoginResponse.java
@@ -1,4 +1,4 @@
-package com.igrowker.wander.dto;
+package com.igrowker.wander.dto.user;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
